+++ frontends/common/src/it/scala/stainless/ExtractionSuite.scala
@@ -1,4 +1,4 @@
-/* Copyright 2009-2016 EPFL, Lausanne */
+/* Copyright 2009-2017 EPFL, Lausanne */
 
 package stainless
 
@@ -6,8 +6,9 @@ import scala.util.{ Try, Success, Failure }
 
 import org.scalatest._
 
-class ExtractionSuite extends FunSpec with inox.ResourceUtils with InputUtils {
-
+/** Subclass are only meant to call [[testExtractAll]] and [[testRejectAll]] on
+ *  the relevant directories. */
+abstract class ExtractionSuite extends FunSpec with inox.ResourceUtils with InputUtils {
 
   private def testSetUp(dir: String): (inox.TestSilentReporter, inox.Context, List[String]) = {
     val reporter = new inox.TestSilentReporter
@@ -62,15 +63,25 @@ class ExtractionSuite extends FunSpec with inox.ResourceUtils with InputUtils {
     }
   }
 
+  // Tests that programs are rejected either through the extractor or through
+  // the TreeSanitizer.
   def testRejectAll(dir: String): Unit = {
     val (reporter, ctx, files) = testSetUp(dir)
 
     describe(s"Programs extraction in $dir") {
-      val tryPrograms = files map { f => f -> Try(loadFiles(ctx, List(f))._2) }
+      val tryPrograms = files map { f =>
+        f -> Try {
+          val program = loadFiles(ctx, List(f))._2
+          extraction.TreeSanitizer.check(program)
+          program
+        }
+      }
+
       it("should fail") {
         tryPrograms foreach { case (f, tp) => tp match {
           // we expect a specific kind of exception:
           case Failure(e: stainless.frontend.UnsupportedCodeException) => assert(true)
+          case Failure(e: stainless.extraction.MissformedStainlessCode) => assert(true)
           case Failure(e) => assert(false, s"$f was rejected with $e")
           case Success(_) => assert(false, s"$f was successfully extracted")
         }}
@@ -78,13 +89,6 @@ class ExtractionSuite extends FunSpec with inox.ResourceUtils with InputUtils {
     }
   }
 
-  testExtractAll("verification/valid")
-  testExtractAll("verification/invalid")
-  testExtractAll("verification/unchecked")
-  testExtractAll("imperative/valid")
-  testExtractAll("imperative/invalid")
-  testExtractAll("termination/valid")
-  testExtractAll("termination/looping")
-  testExtractAll("extraction/valid")
-  testRejectAll("extraction/invalid")
 }
+
+
