+++ frontends/scalac/src/it/scala/stainless/ScalacExtractionSuite.scala
@@ -1,0 +1,11 @@
+/* Copyright 2009-2017 EPFL, Lausanne */
+
+package stainless
+
+class ScalacExtractionSuite extends ExtractionSuite {
+
+  testExtractAll("extraction/valid")
+  testRejectAll("extraction/invalid")
+
+}
+
